requirements.txt
ttkbootstrap
pillow
pyperclip
